refactor(creature): split MonsterFormatExt into CreatureFormatExt and MonsterFormatExt
